Created basic databinding structures for adapters

+++ app/src/main/java/com/example/alsess/model/ApiProductsModel.kt
@@ -11,7 +11,7 @@ data class ApiProductsModel(
 )
 
 data class Rating(
-    val rate: Double,
+    val rate: Float,
     val count: Int,
 )
